create script and supporting files, based on another script as template
Used another decaf script that I made as the starting point of the project, and then just modified all the files to what I need. It contains everything except for the CI files, but I'm going to add those later with a pull request so we can test that it all works.